Moving error check into storage constructor
