+++ packages/bitcore-client/lib/storage.js
@@ -4,6 +4,8 @@ const levelup = require('levelup');
 const leveldown = require('leveldown');
 const Encrypter = require('./encryption');
 const bitcoreLib = require('bitcore-lib');
+const util = require('util');
+const accessAsync = util.promisify(fs.access);
 
 class Storage {
   constructor(params) {
@@ -20,6 +22,15 @@ class Storage {
       }
     }
     this.path = path || `${basePath}/bitcoreWallet`;
+    if(!createIfMissing) {
+      try {
+        await accessAsync(path, fs.constants.F_OK | fs.constants.R_OK);
+        await accessAsync(path + '/LOCK' || path + 'LOCK', fs.constants.F_OK | fs.constants.R_OK | fs.constants.W_OK);
+        await accessAsync(path + '/LOG' || path + 'LOG', fs.constants.F_OK | fs.constants.R_OK | fs.constants.W_OK);
+      } catch (err) {
+        throw new Error('Invalid wallet path');
+      }
+    }
     this.db = levelup(leveldown(this.path), { createIfMissing, errorIfExists });
   }
   async loadWallet(params) {
